client cmd-line interface entry point
