+++ telnetlib3/accessories.py
@@ -1,6 +1,7 @@
 """Accessory functions."""
 
-__all__ = ('name_unicode', 'eightbits')
+__all__ = ('name_unicode', 'eightbits', 'make_logger', 'get_encoding')
+
 
 def name_unicode(ucs):
     """ Return 7-bit ascii printable of any string. """
@@ -18,6 +19,7 @@ def name_unicode(ucs):
         rep = r'\x{:02x}'.format(bits)
     return rep
 
+
 def eightbits(number):
     """
     Binary representation of ``number`` padded to 8 bits.
@@ -30,3 +32,28 @@ def eightbits(number):
     # useful only so far in context of a forwardmask or any bitmask.
     prefix, value = bin(number).split('b')
     return '0b%0.8i' % (int(value),)
+
+
+def make_logger(loglevel='info', logfile=None):
+    import logging
+    fmt = '%(asctime)s %(levelname)s %(filename)s:%(lineno)d %(message)s'
+    lvl = getattr(logging, loglevel.upper())
+    logging.getLogger().setLevel(lvl)
+
+    _cfg = {'format': fmt}
+    if logfile:
+        _cfg['filename'] = logfile
+    logging.basicConfig(**_cfg)
+
+    return logging.getLogger(__name__)
+
+
+def get_encoding():
+    import locale
+    import codecs
+    locale.setlocale(locale.LC_ALL, '')
+    return codecs.lookup(locale.getpreferredencoding()).name
+
+
+def repr_mapping(mapping):
+    return ' '.join('='.join(map(str, kv)) for kv in mapping.items())
