+++ tox.ini
@@ -5,6 +5,7 @@ skip_missing_interpreters = true
 [testenv]
 whitelist_externals = cp
 setenv = PYTHONIOENCODING=UTF8
+         PYTHONASYNCIODEBUG=1
 deps = -rrequirements-tests.txt
 install_command = pip install --upgrade {packages}
 usedevelop = True
